Seperates tests for different modules into different files
