+++ tests/test_integrations.py
@@ -1,35 +1,17 @@
 #!/usr/bin/env python3.7
 # -*- coding: UTF-8 -*-
 
-"""Test that SNData interfaces with SNCosmo correctly.
-Tests are run using csp.dr3
+"""Test that SNData interfaces correctly with each integrated service. Tests
+are run using data from CSP DR3.
 """
 
 from unittest import TestCase
 from warnings import warn
 
-import sncosmo
 from requests.exceptions import HTTPError
 
-from sndata import csp, query_ned_coords, query_osc, query_osc_photometry, \
+from sndata import query_ned_coords, query_osc, query_osc_photometry, \
     query_osc_spectra
-
-
-class SNCosmo(TestCase):
-    """Tests for integrated SNCosmo services"""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.module = csp.dr3
-        cls.module.download_module_data()
-
-    def test_filter_registration(self):
-        """Test csp registered bands do not have empty names"""
-
-        self.module.register_filters()
-        for band_name in self.module.band_names:
-            sncosmo_band = sncosmo.get_bandpass(band_name)
-            self.assertEqual(band_name, sncosmo_band.name)
 
 
 class NED(TestCase):
@@ -57,17 +39,17 @@ class OSC(TestCase):
     def setUpClass(cls):
         cls.test_id = '2011fe'
 
-    def test_general_query(self):
+    def test_general_query_not_empty(self):
         """Test queries for an object's metadata"""
 
         self.assertTrue(query_osc(self.test_id))
 
-    def test_photometry_query(self):
+    def test_photometry_query_not_empty(self):
         """Test queries for an object's photometric data"""
 
         self.assertTrue(query_osc_photometry(self.test_id))
 
-    def test_spectra_query(self):
+    def test_spectra_query_not_empty(self):
         """Test queries for an object's spectroscopic data"""
 
         self.assertTrue(query_osc_spectra(self.test_id))
